requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
aiohttp<4,>=3.10
uvloop==0.19.0
asyncpg==0.29.0
